python/requirements.txt
requests>=2.32.0,<3
Pillow>=10,<13
